SPR1-813 Retry socket timeouts
The S3 chunk store has a custom HTTP response reader (`_read_chunk`)
for performance reasons. Unfortunately this bypasses the urllib3 retry
machinery, also for read timeouts wile streaming the response content.

We are already doing our own retries for status errors, so extend this
to deal with read timeouts too. The basic idea is to catch the socket
timeout (the old `socket.timeout` or newer `TimeoutError` since 3.10)
and turn it into an urllib3 `ReadTimeoutError` to trigger a retry.

The most important bit is that we now also test this :-) We use token
proxy suggestions to pause the reading mid-stream for a specified
period.

Report the number of retries when the request finally fails, as an
assurance that retries actually happened.

This addresses #362 and TESD-58 and maybe even SPR1-368.