SPR1-813 Test the case of a truncated RDB file
When testing the token S3 server, add a suggestion to truncate the
RDB file after 1000 bytes. (For the record, the RDB file is 4719 bytes
long.) This triggers the exact error reported 2 years ago to the day
in SPR1-813. The test fails with::

  "katsdptelstate.errors.RdbParseError: Invalid RDB file object"

As the ticket said: "Think of some unit tests to elicit this behaviour."