SPR1-813 Early resets are connection errors
When the connection is reset before anything is returned, urllib3
and/or requests treat it as a connection error instead of a read error.
If the reset occurs after some data has been returned, it is a read
error.

Since this is a grey area anyway (see e.g. psf/requests#2230 and
urllib3/urllib3#470), produce the simplest code and worry about the
semantics later. The S3 chunk store is safe to retry under most
circumstances.

This updates the unit test so long to detect a connection error, but
it still fails.