SPR1-813 Construct timeouts and retries better
We are working towards the ability to override both timeouts and retries
per request. This means that we have to construct or normalise
a timeout tuple and `Retry` object in multiple places. Add two helper
functions to build and sanitise these objects. This also makes the
__init__ and request code more understandable.

In the process, fix the docstring for `timeout`. We accept `None`
instead of a float and it means to wait forever (no timeout).
This happens to be the default in Requests, so it will "leave things
unchanged", but that description is misleading otherwise.