SPR1-813 Restore _raise_for_status
In preparation for retrying reset connections, restore the standalone
`_raise_for_status` function, after removing it 3 years ago (with
apologies to 1e4404bc :-) ). Back then the concern was that it will
need access to more attributes of the store, but now we simply pass
those attributes -- and throw in `retries` to boot. The main benefit
is a cleaner `S3ChunkStore.request` function, which promotes insight
into the looming retry rework.

The `complete_request` method has no need for `chunk_name` anymore,
so absorb it into the kwargs. Change `chunk_name` from positional
to keyword-only in the process.