SPR1-813 Don't do all retries ourselves
In an epic about-turn, do the connect and status retries inside urllib3,
and only the response read retries on the outside. This avoids having
to reinvent the retry wheel (e.g. redirects??). Status retries can
finally go back into urllib3 because we got rid of the 555 fake status
code.

The trick is to override `max_retries` on the HTTPAdapter of the
Requests Session. Use the same `Retry` object throughout, retrieving it
from the response after each request, incrementing it on response errors
and passing it back to the adapter.

To simplify the code, introduce a standalone `_request` function that
catches socket timeouts and reset connections on the response and turns
it into the appropriate urllib3 exceptions. This replaces the quirky
`S3ChunkStore._request`.

When the retries run out, catch either urllib3's `MaxRetryError` (raised
while reading the response) or the appropriate Requests exception. Map
these to S3ServerGlitch using the error map, for simplicity. A possible
caveat is `requests.exceptions.ConnectionError`, which now becomes
`StoreUnavailable` but this might be triggered by read failures (e.g.
too many ProtocolErrors while reading the headers).

Let the RDB reader use `complete_request`. This still fails in the
token-based unit test, because IncompleteReads are not handled directly
yet and TruncatedRead only applies to `read_array`.

We now want the default `raise_on_status=True` since urllib3 is handling
those retries.