SPR1-813 Rework _DetectTruncation class
This now raises urllib3's `IncompleteRead` instead of creating its
own version (`TruncatedRead`). It also wraps `readinto` just like it
did for `read`, to be more generic.

Part of the fun is figuring out how many bytes were read up to this
point, and how many bytes are left in the response (more useful than
the requested bytes in the actual read / readinto call). If we are at
the urllib3 level, this is simple. The faster (and lower) httplib level
does not track the bytes read and is unseekable, so track it ourselves
via `_content_length`.

Turn IncompleteReads into ProtocolErrors in `S3ChunkStore._request`,
so that it may trigger read retries.