SPR1-813 Finally fix glitches on RDB file reads
RDB files are large (several MB) and things can go wrong during its
retrieval. Extend the read retry mechanism to it too. Check for short
reads after downloading the content and raise an IncompleteRead before
it hits `telstate.load_from_file` and turns into an `RdbParseError`.

Extend the test setup so that the token also allows paths that start
with suggestions. This is necessary because the RDB URL is used to
find the chunk store URL, by going one level up.

For the `test_rdb_support` test we now have::

RDB: /please-truncate-read-after-1000-bytes-for-0.4-seconds/1234567890/1234567890_sdp_l0.rdb
store: /please-truncate-read-after-1000-bytes-for-0.4-seconds/

If the suggestion was in the middle of the RDB like before, it breaks
because the chunk store will be located inside the RDB bucket::

RDB: /1234567890/please-truncate-read-after-1000-bytes-for-0.4-seconds/1234567890_sdp_l0.rdb
store: /1234567890/

A side effect of this is that not just the RDB file, but all chunks
larger than 1000 bytes will be dutifully truncated and recovered, since
the suggestion appear in their URLs too :-)

This finally addresses @IanHeywood's comment in #362 two weeks ago
that sparked this whole PR...