SPR1-813 Let RDB files retry on reset connections
RDB files are loaded as generic Ceph objects using `_read_object`
and a non-streaming request. This actually preloads the response
content as part of `Session.request` (see the end of `Session.send`).
The ConnectionResetError therefore happens even before the response
is returned.

The solution is to put the `session.request` with-statement inside
the `try...except` block in `S3ChunkStore._request`, so that we can
catch glitches inside the request itself.

The second part of the solution is to mangle the exceptions back to
urllib3's `ProtocolError`. When accessing the response `content`
property, ConnectionResetErrors will become ProtocolErrors, only to
be turned into ChunkedEncodingErrors by Requests. Revert that back
to a ProtocolError in order to trigger retries in the calling context.

Expand the tests to check what happens when RDB loading finally fails,
and to check whether generic objects can survive some basic glitches.

For the record, I tried to use urllib3's `enforce_content_length`
property on `HTTPResponse`, but had no real luck with it in both
streaming and non-streaming cases, so leave the _DetectTruncation
machinery in place.