SPR1-813 Explicitly set NPY version 2 in tests
We want to test NPY files with version 2.0 headers, but the `np.save`
function picks version 1.0 by default unless your header size actually
exceeds 64 KiB. The relevant test therefore creates a 70 KB header
to force version 2.0 files.

NumPy >= 1.23.5 now complains if the header size is larger than 10 KB.
It is considered a risk to pass that to `literal_eval`. One way around
it is to increase the threshold via the `max_header_size` parameter
to the header reader functions, but that only exists on recent NumPy
versions. The crash is also avoided by allowing pickles, but that is
not appealing either.

The solution is to use the lower-level `write_array` function, which
allows us to set the header version explicitly. This also removes the
need to suppress a warning about version 2 compatibility.