SPR1-2655 Find original chunks from indexed darray
Since ChunkStore identifies chunks by their slice specs, it is easy to
generate a list of all chunk names in a full dask array, using
`da.core.slices_from_chunks` and `ChunkStore.chunk_metadata`.

It is trickier if the dask array is indexed or sliced, since we want
the original chunk names as they are stored in S3 or on disk.

One trick is to leverage `PlaceholderChunk`. Extend it to store the
original chunk name. Even if the dummy chunk is sliced or indexed,
the name is preserved. And dummy chunks that are discarded will drop
the corresponding chunk names too.

Then add a new "error" mode called "dryrun", which pretends that all
chunks are lost and replaced by `PlaceholderChunk`s. This avoids any
data transfers in the process. The resulting dask array can't be
compute()ed directly (since placeholders can't be concatenated) but
the individual chunks in the block view can. The resulting dummy
chunks can be inspected to find the original chunk identities.