SPR1-2655 Restore --workers script argument
This now correctly sets --transfers (and --checkers), but don't go
too wild.

While we are at it, avoid HEAD requests to read the modification time
by setting --checksum (which also doublechecks that the data is good).
This will slow things down a little bit on a resumed copy but is safe.
Also avoid GET requests to read directory listings and use fast
ListObjectsV2 instead via --fast-list. Both recommendations are from
the S3-specific rclone docs (https://rclone.org/s3/).

The rclone progress printout bombs out with a large number of threads.
If the transfers and checkers won't fit on the terminal screen, rather
revert to a one-line printout. Hopefully this workaround will be
unnecessary in a future version of rclone.