SPR1-2655 Ensure that transfers can be overridden
The default setting was masking the user-supplied rclone flags because
it was at the end of the list of arguments.

I've also noticed that `--transfers=10` won't be split up by
parse_known_args and will therefore be ignored by
rclone_fit_output_to_terminal. Leave this as is since it is already
a bit of a silly feature.