SPR1-2912 Test scan/compscan/target selection
Use pytest parametrization to feed various inputs to `DataSet.select`
and compare the resulting selected dumps to the expected ones.

This uncovered a minor niggle: you can't use the shortened and more
readable "target, radec, 19:39, -63:42", because the Target object will
have a description string of "target, radec, 19:39:00.00, -63:42:00.0"
instead and the comparison will fail.

Move the `test_selecting_antenna` method to the new test class too,
since it was a fish out of water in the `TestVirtualSensors` class.