SPR1-2912 Select multiple targets with same name
A katpoint Catalogue can have multiple different targets with the same
name, usually differing in tags only. The DataSet selection mechanism
still used a dict-style lookup for target names though, which is
constrained to return a single target: the last target with the given
name.

This would normally not be a big issue as there are other ways to select
targets. It became a problem after #269 (released in katdal 0.18 in
April 2021) which added field selection to mvftoms.py. This now always
selects targets by name, even if the user wants all the targets and the
selection is therefore unnecessary. This caused folks to lose e.g. a
flux or bpcal calibrator in the MS conversion if they used the same
target as a gain calibrator later on in the observation.

Use a private function (for now) to access the full list of targets
with a given name, and add all of them to the selection. We should
promote Catalogue._targets_with_name to the main katpoint API in its
1.0 release.