SPR1-2990 Store URL creation is generic and public
Rename the `_chunk_url` private method to `make_url` so that it is more
generic and also part of the public interface. The chunk extension is
now left up to the caller. The rationale is that we want to concentrate
all URL assembly in a single function, where we can also fix problems
with the bucket name.