SPR1-3007 Restore np.bool_ for Numba function
The Numba @jit(nopython=True) option is very particular about types.
It does not like `bool` or `np.bool` (since `np.bool is bool` anyway)
but prefers NumPy's custom `np.bool_` type. The error message mentions

  numba.errors.TypingError: Failed in nopython mode pipeline
  (step: nopython frontend)
  Invalid use of Function(<built-in function empty>) with argument(s)
  of type(s): (Literal[int](128), dtype=Function(<class 'bool'>))

Revert this back the way it was... It was inadvertently converted
during the great rush to replace np.bool. We really need a unit test
for this function (or better yet, a dask replacement).