SPR1-2997 Fix test failures on Apple Silicon
Some complex arrays previously tested equal on Intel machines but this
is an coincidence, given its dependence on the order of floating-point
operations. Differences crept in on arm64 / Apple Silicon processors.
Use the new assert_array_equal_within_n_ulps to bound these errors
in a sensible way. The thresholds were set so that 100 repetitions of
each test passes.