SPR1-2655 Create underlying storage arrays better
In order to figure out the chunks associated with a selection, we need
a consistent set of vis/flags/weights arrays indexed by a single set
of boolean arrays. This is complicated by the fact that L0 and L1 shapes
may differ due to different chunking schemes, a problem encountered
in #378.

Rather use the newly exposed `chunk_info` member on `d.source.data` to
get a consistent and normalised set of underlying vis/flags/weights
arrays, as the shape alignment is already taken care of by the time that
the `ChunkStoreVisFlagsWeights` object is constructed. This info is hard
to extract from the final massive dask graph. Our feeble attempt to load
the original arrays directly from the telstate `chunk_info` was brittle.

Throw in `preselect_index` to boot, so that the script will also support
preselection if someone goes there.

This addresses #378.