SPR1-3236 Don't test 0-D chunks in 3.12 S3 store
Before Python 3.12 a zero-dimensional memoryview had a len of 1. Since
Python 3.12 it raises `TypeError` instead. This prevents
`S3ChunkStore.put_chunk` from storing zero-dimensional arrays, since
deep in the bowels of `urllib3` it tries to chunk up the memoryview and
then casts it to `bool` via `if chunk`, which triggers `len(chunk)`.

The support for zero-dimensional chunks is only dropped on Python 3.12+
for `S3ChunkStore`. I'm keeping it around for the rest because it was
useful in the past to test edge cases. On the other hand, we have never
stored 0-dim chunks anyway, so the whole thing is a bit moot. :-)