SPR1-3236 Fix tests to use float64 timestamps
NumPy 2.0 changed the data type promotion rules so that e.g.
`np.float32(1) + 1` is `np.float32(2.0)` instead of `np.float64(2.0)`.
This exposed a quirk in the unit tests where the expected timestamps
where generated using an array of `np.float32` (for unknown reasons).
When this array was added to the `float64` first timestamp the old
promotion rules corrected it to `float64`, but not anymore.

I can't figure out why it was float32 in the first place, but I have
now removed the dtype and centralised the parameter values as constants
for consistency.