SPR1-3236 Work around Dask slicing bug for now
The combination of `np.newaxis` and fancy indexing in the same index
has been giving problems in Dask since 2024.8.0 (see dask/dask#11993).
Skip the relevant tests for affected Dask versions. I have not seen a
report of these errors in actual use. Hopefully before that we'll be
able to recommend a fixed Dask version.