SPR1-3236 Suppress NaN warnings in np.reciprocal
When the input to `np.reciprocal` contains 0j, or real/imag parts that
are `np.nan`, the function call emits the following warning:

  RuntimeWarning: invalid value encountered in reciprocal

The output is `np.nan + 1j * np.nan` in both cases, which is what we
want: an INVALID_GAIN that will be flagged. The warning really serves
no purpose except to clog up cal pipeline logs and pytest output.

Ignore the warning via the usual filtering mechanism. An alternative
would have been to avoid inverting invalid values like this:
```python
def complex_reciprocal(x):
    """Invert `x` but not if its magnitude is 0.0 or NaN."""
    invalid = (x == 0) | np.isnan(x)
    out = np.full_like(x, INVALID_GAIN)
    np.reciprocal(x, out=out, where=~invalid)
    return out
```
This has the added complication of imposing a complex64 output type and
being twice as slow as the warning filter. That's why we picked the
former option.